1st project - Number guess
A program that generates that generates a random number and asks the player to guess it . If the player guess the higher number than the actual the program displays "Lower number please". Similarly if the guess is too low the program prints "Higher number please". When player guesses the correct number, the program displays the number of guesses the player used to arrive at the number.
No of loop used = 08